requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1

sqlalchemy==2.0.25
pydantic==2.6.1

python-dotenv==1.0.1
python-multipart==0.0.9

openai==1.12.0
httpx==0.26.0
tiktoken==0.6.0
numpy
hnswlib==0.8.0
pdfplumber==0.11.0
